backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.25.0
sqlmodel==0.0.18
psycopg2-binary>=2.9.9
python-dotenv==1.0.1
pydantic==2.8.0
pydantic-core==2.23.0
python-multipart==0.0.10
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography>=42.0.0
openai>=1.0.0